refactor: move BlogPost model into separate file
