+++ models/index.js
@@ -1,28 +1,20 @@
-const { DataTypes } = require("sequelize");
 const sequelize = require("../config/database");
+const User = require("./user");
+const BlogPost = require("./blogPost");
 
-// Define BlogPost model
-const BlogPost = sequelize.define(
-  "BlogPost",
-  {
-    title: {
-      type: DataTypes.STRING(100),
-      allowNull: false,
-    },
-    content: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-    },
-    author: {
-      type: DataTypes.STRING(100),
-      allowNull: false,
-    },
-  },
-  {
-    timestamps: true,
-    createdAt: "created_at",
-    updatedAt: false,
-  },
-);
+// Sync models with the database
+const initialiseModels = async () => {
+  try {
+    await sequelize.sync();
+    console.log("Database synchronised successfully.");
+  } catch (error) {
+    console.error("Error syncing the database:", error);
+  }
+};
 
-module.exports = { sequelize, BlogPost };
+module.exports = {
+  sequelize,
+  User,
+  BlogPost,
+  initialiseModels,
+};
